Fixed incorrect handling in 1.18(1.18.1) and 1.18.2

+++ MinecraftClient/McClient.cs
@@ -2490,7 +2490,7 @@ namespace MinecraftClient
         /// <param name="action">0 to create/update an item. 1 to remove an item.</param>
         /// <param name="objectivename">The name of the objective the score belongs to</param>
         /// <param name="value">he score to be displayed next to the entry. Only sent when Action does not equal 1.</param>
-        public void OnUpdateScore(string entityname, byte action, string objectivename, int value)
+        public void OnUpdateScore(string entityname, int action, string objectivename, int value)
         {
             DispatchBotEvent(bot => bot.OnUpdateScore(entityname, action, objectivename, value));
         }
